First pass at fixing Abe's comments.
