+++ analysis/paladinholy/src/CombatLogParser.ts
@@ -29,8 +29,8 @@ import CastBehavior from './modules/features/CastBehavior';
 import CooldownThroughputTracker from './modules/features/CooldownThroughputTracker';
 import MasteryEffectiveness from './modules/features/MasteryEffectiveness';
 import Overhealing from './modules/features/Overhealing';
-import UntemperedDedication from './modules/items/UntemperedDedication';
 import MightOfTheMountain from './modules/racials/MightOfTheMountain';
+import UntemperedDedication from './modules/shadowlands/conduits/UntemperedDedication';
 import MaraadsCastRatio from './modules/shadowlands/legendaries/MaraadsCastRatio';
 import MaraadsOverheal from './modules/shadowlands/legendaries/MaraadsOverheal';
 import ShockBarrier from './modules/shadowlands/legendaries/ShockBarrier';
@@ -111,7 +111,7 @@ class CombatLogParser extends CoreCombatLogParser {
     ashenHallow: AshenHallow,
     divineToll: DivineToll,
 
-    // Items
+    // Conduits
     UntemperedDedication: UntemperedDedication,
   };
 }
